styling
decided to work on the styling and just removed the update entirely since after rereading everything it doesn't seem to be required - I just assumed that since it was one of the CRUD action types in the comments in the index file that we had to figure that out but the README doesn't say anything about it so I'll tinker with that later. Added a table for displaying my smurfs.